backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.135.3",
    "google-genai>=1.72.0",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.44.0",
    "websockets>=16.0",
]
